added hyperstructure warning
