+++ client/src/ui/components/hyperstructures/HyperstructureDetails.tsx
@@ -1,11 +1,22 @@
+import { LeaderboardManager } from "@/dojo/modelManager/LeaderboardManager";
 import { Tabs } from "@/ui/elements/tab";
+import { ID } from "@bibliothecadao/eternum";
 import { useMemo, useState } from "react";
 import { CoOwners } from "./CoOwners";
 import { Leaderboard } from "./Leaderboard";
-import { ID } from "@bibliothecadao/eternum";
 
 export const HyperstructureDetails = ({ hyperstructureEntityId }: { hyperstructureEntityId: ID }) => {
   const [selectedTab, setSelectedTab] = useState(0);
+
+  const coOwnersWithTimestamp = useMemo(() => {
+    const latestChangeEvent = LeaderboardManager.instance().getCurrentCoOwners(hyperstructureEntityId);
+    if (!latestChangeEvent) return undefined;
+
+    const coOwners = latestChangeEvent.coOwners;
+    const timestamp = latestChangeEvent.timestamp;
+
+    return { coOwners, timestamp };
+  }, [hyperstructureEntityId]);
 
   const tabs = useMemo(
     () => [
@@ -23,12 +34,15 @@ export const HyperstructureDetails = ({ hyperstructureEntityId }: { hyperstructu
         label: (
           <div className="flex relative group flex-col items-center">
             <div>Co Owners</div>
+            {!coOwnersWithTimestamp && <div className="text-xs text-red animate-pulse">Required for points!</div>}
           </div>
         ),
-        component: <CoOwners hyperstructureEntityId={hyperstructureEntityId} />,
+        component: (
+          <CoOwners hyperstructureEntityId={hyperstructureEntityId} coOwnersWithTimestamp={coOwnersWithTimestamp} />
+        ),
       },
     ],
-    [hyperstructureEntityId],
+    [hyperstructureEntityId, coOwnersWithTimestamp],
   );
 
   return (
