+++ eternum-docs/docs/pages/mechanics/hyperstructures.mdx
@@ -56,11 +56,20 @@ all contributors Points are calculated and distributed automatically upon constr
 
 ### Holding shares of the Hyperstructure
 
-Once operational, Hyperstructures become ongoing sources of points:
+⚠️ **IMPORTANT: Setting Co-owners is Required for Points** ⚠️
+
+Once a Hyperstructure is operational, you MUST set the co-owners and their share allocations before any points can be
+generated. If you don't set co-owners, NO POINTS WILL BE ACCRUED, even if the Hyperstructure is complete. You can:
+
+- Assign 100% of shares to yourself
+- Distribute shares among multiple players
+- Modify share allocations (with a 48-hour cooldown)
+
+Once co-owners are set, points are generated as follows:
 
 - {EternumGlobalConfig.hyperstructures.hyperstructurePointsPerCycle} points are generated each Eternum cycle
-- Distribution follows a share-based system
-- The structure owner can control and set share allocations
+- Points are automatically distributed according to share percentages
+- The structure owner controls share allocations
 - Share modifications have a 48-hour cooldown period
 
 ## Strategic Considerations
@@ -72,3 +81,4 @@ Once operational, Hyperstructures become ongoing sources of points:
 > - Defend completed structures to maintain control
 > - New owner gains share control upon capture
 > - Previous share settings reset on capture
+> - **Remember to set co-owners immediately after construction to start earning points**
